refactor: NicknameRepository 기능을 UserRepository로 통합
- NicknameRepository 및 NicknameRepositoryImpl 삭제
- UserRepository 인터페이스에 nickname 관련 메서드(save, get, observe) 추가
- UserRepositoryImpl에서 UserApi 및 ProfileDataStore를 사용하여 nickname 기능 구현
- RepositoryModule에서 NicknameRepository 바인딩 제거 및 AuthRepository에 @Singleton 추가
