+++ data/src/main/java/com/smtm/pickle/data/di/RepositoryModule.kt
@@ -2,11 +2,9 @@ package com.smtm.pickle.data.di
 
 import com.smtm.pickle.data.repository.AuthRepositoryImpl
 import com.smtm.pickle.data.repository.LedgerRepositoryImpl
-import com.smtm.pickle.data.repository.NicknameRepositoryImpl
 import com.smtm.pickle.data.repository.UserRepositoryImpl
 import com.smtm.pickle.domain.repository.AuthRepository
 import com.smtm.pickle.domain.repository.LedgerRepository
-import com.smtm.pickle.domain.repository.NicknameRepository
 import com.smtm.pickle.domain.repository.UserRepository
 import dagger.Binds
 import dagger.Module
@@ -19,15 +17,12 @@ import javax.inject.Singleton
 abstract class RepositoryModule {
 
     @Binds
+    @Singleton
     abstract fun bindAuthRepository(authRepositoryImpl: AuthRepositoryImpl): AuthRepository
 
     @Binds
     @Singleton
     abstract fun bindUserRepository(impl: UserRepositoryImpl): UserRepository
-
-    @Binds
-    @Singleton
-    abstract fun bindNicknameRepository(impl: NicknameRepositoryImpl): NicknameRepository
 
     @Binds
     @Singleton
